Restructure lib src files
